- Getting affymatrix data is now running and saving the tsv with the data quality scores

+++ GettingData/getting_affymatrix_data.R
@@ -146,16 +146,25 @@ target_geo_ids <- c( # these are all the human GSE's (hs)
 #------------------functions----------------------
 
 quality_control_removal <- function(cel_dir_path){
-  cel_file_paths = list.celfiles(cel_dir_path, listGzipped = TRUE, full.name = TRUE)
+  threshold = 0.15
+  
+  cel_file_paths = list.celfiles(cel_dir_path, full.name = TRUE)
   cel_files = read.celfiles(cel_file_paths)
   test_results = arrayQualityMetrics(expressionset = cel_files, force = TRUE, outdir = "quality_output_file")
   unlink("quality_output_file", recursive = TRUE)
-  which_integers = test_results$modules$maplot@outliers@which
+  statistic_list = test_results$modules$maplot@outliers@statistic
+  statistic_tibble = as_tibble(statistic_list)
+  statistic_tibble = add_column(statistic_tibble, cel_file = cel_file_paths)
+  
+  write_tsv(statistic_tibble, "Data/Affymetrix/quality_output_file.tsv", append = TRUE, col_names = FALSE)
+  
   # Here we need to go into the file and delete the files based off the integers that are returned probably in a for loop
-  for (num in which_integers){
-    file_for_delete = cel_file_paths[num]
-    print(paste0("Deleting ", file_for_delete))
-    file.remove(file_for_delete)
+  for (row in 1:nrow(statistic_tibble)){
+    
+    if (statistic_tibble[row, "value"] > threshold){
+      file_for_delete = cel_file_paths[row]
+      file.remove(file_for_delete)
+    }
   }
 }
 
@@ -260,10 +269,10 @@ get_scan_upc_files <- function(geo_id, platform_to_package_list){
   quality_control_removal(tar_file_output_f)
   platform = platform_list[[geo_id]]
   pkgName = platform_to_package_list[[platform]]
-
+  
   # last step to converting the information
   normalized = SCAN(celFilePattern, convThreshold = .9, probeLevelOutDirPath = NA, probeSummaryPackage=pkgName)
-
+  
   return (normalized)
   
 }
@@ -312,6 +321,7 @@ print(paste('Total time: ', format_time_diff(total_time)))
 # 
 # data <- read_tsv(gzfile("Data/Affymetrix/GSE138861.tsv.gz"))
 # view(data)
+
